Coordinator/requirements.txt
requests>=2.26.0
google-cloud-storage>=2.1.0